Use TypeConv to obtain Param Writers
The previous system, based on Obtain Writer.vi, relied on hard-coded
type names or incomplete heuristics to determine the Param Writer
subclass to use. This was not scalable.

The new system uses TypeConv.vi registration, which is extensible to
support new types and categories. It also uses the JSON inputs to
determine the Param Writer subclass to use.