[LabVIEW] Rename "QObject Writer" -> "Identity Writer"
Currently, everything done by this class applies to all Identity
objects. There is nothing QObject-specific (yet). If/When QObject-
specific features are needed in the future, a new QObject Writer class
should be created, which subclasses Identity Writer.

In addition to renaming, this class has been simplified as there is no
need for external callers to retrieve the LVClass reference:
- "Read QObject Class Ref.vi" is removed
- The reference is no longer accessible as a property.