[LabVIEW] Move lq(Set)Manifestation.vi from Community to Internal
Ideally, we'd prevent users from calling these VIs. However, putting
them in Community scope requires LQIdentity and LQSerialData to add all
subclasses to their Friends list. This makes it unfeasible to add
modules in the future, as that would require modifying the core module
to add new Friends.

Now, we rely on the gray VI colour as a visual hint to tell users not to
call these VIs.