[LabVIEW] Clone template VIs into an external temporary directory
Previously, we used the VIs that were copied into the target directory
when the original template LVLib is cloned. If we're to support add-on
modules, then the template LVLib would only be cloned for the kernel
(LQWidgetEngine et al.). Module generation would not involve this
cloning, so we can't use those auto-copied VIs anymore.

To complicate matters, we cannot pass the VIs in the original template
LVLib into the "Create New VI" node, because their subVIs are linked to
the classes in the original template LVLib, not the target LVLib. So,
we need manually copy those VIs and re-link the subVIs to the correct
LVLib.

- Create Temporary Template Directory.vi takes care of the copying and
  re-linking.
- Delete Temporary Template Directory.vi cleans up.
- Generate and Save Method VIs.vi and Create Method VIs.vi now receive
  the path to the directory that holds those temporary VIs, instead of a
  reference to the template LVClass.
- The auto-copied VIs are now deleted earlier, since they're not used
  anymore.