[LabVIEW] Apply inheritances to Projection classes too
Currently, all specified Projections are of the Alias category (as
opposed to the Interface category). For these, the Projection target
class should inherit from the source's ancestor.

This patch unconditionally applies inheritances unconditionally, on the
assumption that all Projections are Aliases. In the future, we should
register the Projection category and apply inheritances selectively.