[LabVIEW] Use LVClass paths instead of LVClass refs where possible
The amount of time needed to create and save a VI is proportional to the
number of VIs already open in memory, it seems. Now, LVClass methods
can't be removed from memory unless the LVClass itself is closed. So, we
want to minimize the number of LVClasses in memory.

Using the same inputs as c82f228b78bf21c053911a590860024ab3074782, this
patch reduces code generation time from 52.3 minutes (Windows 8.1) to
17.2 minutes (Windows 8.1) or 18.0 minutes (Windows 10, same PC).

There is an overhead for reopening LVClasses though, which results in a
minor penalty for very small modules. The time taken to generate Qt SVG
(1 class, 2 VIs) increased from ~10s to ~35s.