Enable connection between all native Qt signals and slots
Qt has many signals and slots that aren't accessible from LabVIEW:
- Those that have multiple parameters
- Those that use types other than the 4 currently supported by LabVIEW
  (bool, int, double, QString)

This patch provides one way to connect with the methods that fit the
above criteria. Even if the connection involves LabVIEW-supported
methods, it's still convenient to create a connection without having to
write LabVIEW code.

Note: This does not work with connecting a signal (native or LabVIEW) to
a LabVIEW-defined signal, because the latter does not have a function
body.