[LabVIEW] Make it safe to call Get Signal Channels.vi at all times
Previously, this VI could only be called to retrieve existing
connections; passing invalid inputs would throw an internal error which
couldn't be caught and handled.

This patch makes the VI output an empty array if the inputs don't match
any existing connections, without throwing an error. This makes it safe
to use this VI to check if a connection exists.