[LabVIEW] Fix signal duplication
LQCoreEngine inserts itself in between the signal and the actual LabVIEW
slot. Even if a signal is connected to multiple slots, there should only
be one connection between the signal and LQCoreEngine, since the latter
takes responsibility for distributing the signal data among the intended
recipients.

This patch adds the necessary checks to avoid duplication.

Fixes bug #47.