Unify string-based connections; handle LabVIEW-defined signals in C++
Introduce LQSignalRelay to act as the "middle man" in signal-signal
connections when the second signal in the pair is LabVIEW-defined. This
provides better performance and stability than the previous design,
which generated two separate LabVIEW User Events per emission, which
could then be processed out-of-order when many signals are emitted
arrive in rapid succession.

In C++, connect_bySignature_qtMethod() is now generalized to handle
LabVIEW-defined signal too. LabVIEW now only needs to pass the method
signatures to C++; LabVIEW no longer needs to care where the signal
was originally defined.