[C++] Avoid accessing dangling LQMetaObjects at shutdown
Internally, ~QApplication() queries the meta object of all QWidgets that
are still around. However, this is a Bad Thing (tm) for widgets that use
dynamic meta objects, because these are all deleted in ~LQApplication()
(which must run before the ~QApplication(), as per C++ rules).

We avoid this potential crash by forcing ~LQApplication() to delete all
remaining widgets before ~QApplication can get to them.