[LabVIEW] Add API to test for null LQIdentity objects
Parts of the Qt API return null pointers to indicate the absence of an
appropriate object (e.g. QWidget::layout()). We want users to be able to
detect such return values, without throwing an error.

The most natural place to put this VI is in LQIdentity.lvclass, although
this means LQIdentity can no longer be treated as an an invisible
internal class.

Closes ticket #43.