Update list of exposed classes/methods/enums
Each method in the collection adds to the overhead of loading a class,
especially if that class is depended upon by many other classes. So, we
need some curation and leave out classes/methods that are unlikely to be
used.

At the same time, we also need more classes/methods for a good initial
demo.

Everything in moderation.