Add license (MPL v2.0)
This license was chosen because it:
- Is copyleft (unlike MIT, BSD-3)
- Allows the library to be used in proprietary software (unlike GPLv3)
- Allows the library to be statically linked to other software without
  imposing restrictions on that software (unlike LGPLv3)

I haven't figured out if enums, data structures, namespaces and class
definitions (i.e. *.ctl, *.lvlib, and *.lvclass files) should have the
copyright notice and license header or not. As an experiment, the text
is added to the files of the code generator, but not to the files of the
actual binding library.