Introduce "lv2014" long-term dev branch
This branch is created to take advantage of the new web-based help
feature introduced in LabVIEW 2014. However, every other feature we use
(so far) are already available in LabVIEW 2013, so we'll refrain from
rolling up the VIs that don't need LabVIEW 2014; the main development
will continue to happen on the master branch using LabVIEW 2013 for now.

We'll periodically merge the master branch into this branch to keep
things in sync.

See ticket #37.