Add support for the LabVIEW Color Box
This involves a new Category, "LQCustomCast", because the data must be modified
at the interface (QRgb interprets the U32 as "AARRGGBB"; LabVIEW interprets the
U32 as "00RRGGBB").

2 new hard-coded types are introduced: "LQRgb" and "LQRgbBrush" (the latter is
for using a LV Color Box with the QBrush constructor).

The main QColor constructor is also replaced by one that takes a LV Color Box.
Alpha values are now handled separately since LabVIEW has no support for it.

Closes ticket #27